Pick medea-reactive changes from the 47-client-state-restoring [run ci]
