+++ crates/medea-reactive/CHANGELOG.md
@@ -13,9 +13,14 @@ All user visible changes to this project will be documented in this file. This p
 
 - `ObservableField` and `ObservableCell` containers ([#81]);
 - `Observable` type alias ([#81]);
-- `OnObservableFieldModification`, `Subscribable` and `Whenable` traits ([#81]).
+- `OnObservableFieldModification`, `Subscribable` and `Whenable` traits ([#81]);
+- Observable collections ([#103]):
+  - `ObservableVec`;
+  - `ObservableHashMap`;
+  - `ObservableHashSet`.
 
 [#81]: /../../pull/81
+[#103]: /../../pull/103
